==> 1_face_non/extended_cleaning_analysis/scripts/4_calculating_zscores_2000.Rmd <==
---
title: "4_calculating_zscores_2000"
author: "Sahar Fatima"
date: "16/09/2020"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

#### note: for faces, 429 = happy, 673 = angry
## load packages 
```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(formattable) # useful for controlling decimal places
library(here)

options(scipen = 999) #set no scientific notation
```

## read in data 

```{r message=FALSE, warning=FALSE}
face <- read_csv(here("extended_cleaning_analysis", "data", "3_face_false_2000.csv")) %>% # face data 
  select(-outlier1.5SD)

non <- read_csv(here("extended_cleaning_analysis", "data", "3_non_false_2000.csv")) %>% # non-face data
  select(-outlier1.5SD)

non %>%
  filter(pp_no == "pp604")
```

## FACE
```{r}
# make data wide, use mutate to make new z variables
face_zscore <- face %>%
  pivot_wider(names_from = muscle, 
              values_from = rms) %>%
  group_by(pp_no) %>%
  mutate(Zbrow = scale(brow, center = TRUE, scale = TRUE)) %>%
  mutate(Zcheek = scale(cheek, center = TRUE, scale = TRUE))
  
summary(face_zscore)
```

## NON-FACE 

```{r}
# make data wide, use mutate to make new z variables
non_zscore <- non %>%
  pivot_wider(names_from = muscle, values_from = rms) %>%
  group_by(pp_no) %>%
  mutate(Zbrow = scale(brow, center = TRUE, scale = TRUE)) %>%
  mutate(Zcheek = scale(cheek, center = TRUE, scale = TRUE))

summary(non_zscore)
```


# Calculating Difference Scores 
need to calculate difference scores from baseline. Need to make BIN wide to allow for calculations across columns. Its a bit difficult to so that for both muscles at the same time, so lets separate and work out the difference scores for brow and cheek separately.

## facial stimuli 

### faces - brow
```{r}
# add a muscle column to make things easier to join back together later
zface_brow <- face_zscore %>%
  mutate(muscle = "brow") %>%
  select(pp_no, condition, bin, trial, muscle, Zbrow) 

# Make the bin column wide and rename bin_0 as BL (i.e. baseline)
zface_brow_wide <- zface_brow %>%
  pivot_wider(names_from = bin, 
              values_from = Zbrow) %>%
  rename(BL = bin_0)

# mutate() diff scores - Use wide columns to calcuate the difference between each bin column and BL, creating a new set of columns starting with “diff”, drop BL column and all columns starting with bin (i.e. raw z scores).

# note - This df contains for each bin the difference between stimulus and basline, so POSITIVE difference scores = greater activity during STIM than BL and NEGATIVE difference scores = greater activity during BL than STIM

zface_brow_diff <- zface_brow_wide %>% 
  mutate(diff_bin1 = bin_1 - BL, 
         diff_bin2 = bin_2 - BL,
         diff_bin3 = bin_3 - BL, 
         diff_bin4 = bin_4 - BL,
         diff_bin5 = bin_5 - BL, 
         diff_bin6 = bin_6 - BL,
         diff_bin7 = bin_7 - BL, 
         diff_bin8 = bin_8 - BL, 
         diff_bin9 = bin_9 - BL, 
         diff_bin10 = bin_10 - BL, 
         diff_bin11 = bin_11 - BL, 
         diff_bin12 = bin_12 - BL,
         diff_bin13 = bin_13 - BL,
         diff_bin14 = bin_14 - BL,
         diff_bin15 = bin_15 - BL,
         diff_bin16 = bin_16 - BL,
         diff_bin17 = bin_17 - BL,
         diff_bin18 = bin_18 - BL,
         diff_bin19 = bin_19 - BL,
         diff_bin20 = bin_20 - BL)%>%
  select(-BL, - starts_with("bin"))

# going to make it long now 
zface_brow_long <- zface_brow_diff %>%
  pivot_longer(names_to = "bin", 
               values_to = "Zdiff", 
               diff_bin1:diff_bin20)
  
```


#### plotting - face brow z-diff 
```{r}
# make conditio a factor 
zface_brow_long$condition <- as.factor(zface_brow_long$condition)

zface_brow_long$bin <- fct_relevel(zface_brow_long$bin, c("diff_bin1", "diff_bin2", "diff_bin3", "diff_bin4", "diff_bin5", "diff_bin6", "diff_bin7", "diff_bin8", "diff_bin9", "diff_bin10", "diff_bin11", "diff_bin12", "diff_bin13", "diff_bin14", "diff_bin15", "diff_bin16", "diff_bin17", "diff_bin18", "diff_bin19", "diff_bin20"))

zface_brow_long %>%
  group_by(condition, bin) %>%
  summarise(meanBROWdiff = mean(Zdiff, na.rm = TRUE)) %>%
  ggplot(aes(x = bin, y = meanBROWdiff, colour = condition, group = condition)) +
  geom_point() + 
  geom_line() +
  labs(title = "brow activity difference from baseline in response to faces") +
  scale_y_continuous(limits = c(-0.2, 0.2))

```


### faces cheek 
```{r}
# add a muscle column to make things easier to join back together later
zface_cheek <- face_zscore %>%
  mutate(muscle = "cheek") %>%
  select(pp_no, condition, bin, trial, muscle, Zcheek) 

# Make the bin column wide and rename bin_0 as BL (i.e. baseline)
zface_cheek_wide <- zface_cheek %>%
  pivot_wider(names_from = bin, 
              values_from = Zcheek) %>%
  rename(BL = bin_0)

# mutate() diff scores - Use wide columns to calcuate the difference between each bin column and BL, creating a new set of columns starting with “diff”, drop BL column and all columns starting with bin (i.e. raw z scores).

# note - This df contains for each bin the difference between stimulus and basline, so POSITIVE difference scores = greater activity during STIM than BL and NEGATIVE difference scores = greater activity during BL than STIM

zface_cheek_diff <- zface_cheek_wide %>% 
  mutate(diff_bin1 = bin_1 - BL, 
         diff_bin2 = bin_2 - BL,
         diff_bin3 = bin_3- BL, 
         diff_bin4 = bin_4 - BL,
         diff_bin5 = bin_5 - BL, 
         diff_bin6 = bin_6 - BL,
         diff_bin7 = bin_7 - BL, 
         diff_bin8 = bin_8 - BL, 
         diff_bin9 = bin_9 - BL, 
         diff_bin10 = bin_10 - BL, 
         diff_bin11 = bin_11 - BL, 
         diff_bin12 = bin_12 - BL,
         diff_bin13 = bin_13 - BL,
         diff_bin14 = bin_14 - BL,
         diff_bin15 = bin_15 - BL,
         diff_bin16 = bin_16 - BL,
         diff_bin17 = bin_17 - BL,
         diff_bin18 = bin_18 - BL,
         diff_bin19 = bin_19 - BL,
         diff_bin20 = bin_20 - BL) %>%
  select(-BL, - starts_with("bin"))

# going to make it long now 
zface_cheek_long <- zface_cheek_diff %>%
  pivot_longer(names_to = "bin", 
               values_to = "Zdiff", 
               diff_bin1:diff_bin20)
  
```

#### plotting - face cheek z-diff 
```{r}
# make condition a factor 
zface_cheek_long$condition <- as.factor(zface_cheek_long$condition)

zface_cheek_long$bin <- fct_relevel(zface_cheek_long$bin, c("diff_bin1", "diff_bin2", "diff_bin3", "diff_bin4", "diff_bin5", "diff_bin6", "diff_bin7", "diff_bin8", "diff_bin9", "diff_bin10", "diff_bin11", "diff_bin12", "diff_bin13", "diff_bin14", "diff_bin15", "diff_bin16", "diff_bin17", "diff_bin18", "diff_bin19", "diff_bin20"))

zface_cheek_long %>%
  group_by(condition, bin) %>%
  summarise(meanCHEEKdiff = mean(Zdiff, na.rm = TRUE)) %>%
  ggplot(aes(x = bin, y = meanCHEEKdiff, colour = condition, group = condition)) +
  geom_point() + 
  geom_line() +
  labs(title = "cheek activity difference from baseline in response to faces") +
  scale_y_continuous(limits = c(-0.2, 0.5))
```



## non-facial sitmuli 
going to calculate z-score differences for non-facial stimuli 

### non-face - brow
```{r}
# add a muscle column to make things easier to join back together later
znon_brow <- non_zscore %>%
  mutate(muscle = "brow") %>%
  select(pp_no, condition, bin, trial, muscle, Zbrow) %>%
  filter(condition != 429)

# Make the bin column wide and rename bin_0 as BL (i.e. baseline)
znon_brow_wide <- znon_brow %>%
  pivot_wider(names_from = bin, 
              values_from = Zbrow) %>%
  rename(BL = bin_0)

# mutate() diff scores - Use wide columns to calcuate the difference between each bin column and BL, creating a new set of columns starting with “diff”, drop BL column and all columns starting with bin (i.e. raw z scores).

# note - This df contains for each bin the difference between stimulus and basline, so POSITIVE difference scores = greater activity during STIM than BL and NEGATIVE difference scores = greater activity during BL than STIM

znon_brow_diff <- znon_brow_wide %>% 
  mutate(diff_bin1 = bin_1 - BL, 
         diff_bin2 = bin_2 - BL,
         diff_bin3 = bin_3 - BL, 
         diff_bin4 = bin_4 - BL,
         diff_bin5 = bin_5 - BL, 
         diff_bin6 = bin_6 - BL,
         diff_bin7 = bin_7 - BL, 
         diff_bin8 = bin_8 - BL, 
         diff_bin9 = bin_9 - BL, 
         diff_bin10 = bin_10 - BL, 
         diff_bin11 = bin_11 - BL, 
         diff_bin12 = bin_12 - BL,
         diff_bin13 = bin_13 - BL,
         diff_bin14 = bin_14 - BL,
         diff_bin15 = bin_15 - BL,
         diff_bin16 = bin_16 - BL,
         diff_bin17 = bin_17 - BL,
         diff_bin18 = bin_18 - BL,
         diff_bin19 = bin_19 - BL,
         diff_bin20 = bin_20 - BL)%>%
  select(-BL, - starts_with("bin"))

# going to make it long now 
znon_brow_long <- znon_brow_diff %>%
  pivot_longer(names_to = "bin", 
               values_to = "Zdiff", 
               diff_bin1:diff_bin20)

znon_brow_long_429 <- znon_brow_long %>%
  filter(condition == 429)
  
```

#### plotting - non brow z-diff 
```{r}
# make conditio a factor 
znon_brow_long$condition <- as.factor(znon_brow_long$condition)

znon_brow_long$bin <- fct_relevel(znon_brow_long$bin, c("diff_bin1", "diff_bin2", "diff_bin3", "diff_bin4", "diff_bin5", "diff_bin6", "diff_bin7", "diff_bin8", "diff_bin9", "diff_bin10", "diff_bin11", "diff_bin12", "diff_bin13", "diff_bin14", "diff_bin15", "diff_bin16", "diff_bin17", "diff_bin18", "diff_bin19", "diff_bin20"))

znon_brow_long %>%
  group_by(condition, bin) %>%
  summarise(meanBROWdiff = mean(Zdiff, na.rm = TRUE)) %>%
  ggplot(aes(x = bin, y = meanBROWdiff, colour = condition, group = condition)) +
  geom_point() + 
  geom_line() +
  labs(title = "brow activity difference from baseline in response to non-facial stimuli") +
  scale_y_continuous(limits = c(-0.2, 0.3))
```


### faces cheek 
```{r}
# add a muscle column to make things easier to join back together later
znon_cheek <- non_zscore %>%
  mutate(muscle = "cheek") %>%
  select(pp_no, condition, bin, trial, muscle, Zcheek) 

# Make the bin column wide and rename bin_0 as BL (i.e. baseline)
znon_cheek_wide <- znon_cheek %>%
  pivot_wider(names_from = bin, 
              values_from = Zcheek) %>%
  rename(BL = bin_0)

# mutate() diff scores - Use wide columns to calcuate the difference between each bin column and BL, creating a new set of columns starting with “diff”, drop BL column and all columns starting with bin (i.e. raw z scores).

# note - This df contains for each bin the difference between stimulus and basline, so POSITIVE difference scores = greater activity during STIM than BL and NEGATIVE difference scores = greater activity during BL than STIM

znon_cheek_diff <- znon_cheek_wide %>% 
  mutate(diff_bin1 = bin_1 - BL, 
         diff_bin2 = bin_2 - BL,
         diff_bin3 = bin_3- BL, 
         diff_bin4 = bin_4 - BL,
         diff_bin5 = bin_5 - BL, 
         diff_bin6 = bin_6 - BL,
         diff_bin7 = bin_7 - BL, 
         diff_bin8 = bin_8 - BL, 
         diff_bin9 = bin_9 - BL, 
         diff_bin10 = bin_10 - BL, 
         diff_bin11 = bin_11 - BL, 
         diff_bin12 = bin_12 - BL,
         diff_bin13 = bin_13 - BL,
         diff_bin14 = bin_14 - BL,
         diff_bin15 = bin_15 - BL,
         diff_bin16 = bin_16 - BL,
         diff_bin17 = bin_17 - BL,
         diff_bin18 = bin_18 - BL,
         diff_bin19 = bin_19 - BL,
         diff_bin20 = bin_20 - BL) %>%
  select(-BL, - starts_with("bin"))

# going to make it long now 
znon_cheek_long <- znon_cheek_diff %>%
  pivot_longer(names_to = "bin", 
               values_to = "Zdiff", 
               diff_bin1:diff_bin20)

z_429 <- znon_cheek_long %>% 
  filter(condition == 429)
  
```


#### plotting - non cheek z-diff 
```{r}
# make condition a factor 
znon_cheek_long$condition <- as.factor(znon_cheek_long$condition)

znon_cheek_long$bin <- fct_relevel(znon_cheek_long$bin, c("diff_bin1", "diff_bin2", "diff_bin3", "diff_bin4", "diff_bin5", "diff_bin6", "diff_bin7", "diff_bin8", "diff_bin9", "diff_bin10"))

znon_cheek_long %>%
  group_by(condition, bin) %>%
  summarise(meanCHEEKdiff = mean(Zdiff, na.rm = TRUE)) %>%
  ggplot(aes(x = bin, y = meanCHEEKdiff, colour = condition, group = condition)) +
  geom_point() + 
  geom_line() +
  labs(title = "cheek activity difference from baseline in response to non-facial stimuli")
```


# combine cheek and brow data for face and non-face, write_csv. 

```{r}
face_zdiff_2000 <- bind_rows(zface_brow_long, zface_cheek_long) %>%
  arrange(pp_no, condition, trial, muscle, bin) %>%
  write.csv(here("extended_cleaning_analysis", "data", "4_face_zdiff_2000.csv"))

non_zdiff_2000 <- bind_rows(znon_brow_long, znon_cheek_long) %>%
  arrange(pp_no, condition, trial, muscle, bin) %>%
  write.csv(here("extended_cleaning_analysis", "data", "4_non_zdiff_2000.csv"))
```
